requirements.txt
aioice==0.9.0
aiortc==1.7.0
altair==5.2.0
anyio==4.2.0
attrs==23.2.0
audio-recorder-streamlit==0.0.8
audioread==3.0.1
av==11.0.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.3
decorator==5.1.1
dnspython==2.6.0
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-crc32c==1.5.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
ifaddr==0.2.0
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
packaging==23.2
pandas==2.2.0
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.0
protobuf==4.25.3
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.1b0
pydub==0.25.1
pyee==11.1.0
Pygments==2.17.2
pylibsrtp==0.10.0
pyOpenSSL==24.0.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
streamlit==1.31.1
streamlit-webrtc==0.47.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.0
tornado==6.4
tqdm==4.66.2
transformers==4.37.2
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
watchdog==4.0.0
zipp==3.17.0
